build fix reported by Rhialto
svn path=/trunk/; revision=243

+++ pan/data/article.cc
@@ -153,7 +153,7 @@ Article :: Part :: set_message_id (const Quark& key_mid, const StringView& mid)
 
   kc = &key.back();
   mc = &mid.back();
-  ke = key.str + 1 + std::max (int(b), (int)minlen - UCHAR_MAX);
+  ke = key.str + 1 + std::max (int(b), (int)minlen - (int)UCHAR_MAX);
   while (kc>ke)
     if (*kc-- != *mc--) { ++kc; break; }
   const size_t e (&key.back() - kc);
